Master and node version default to latest
This is as opposed to letting GKE figure out the latest patch suffix.
This strategy does not work for specifying node version - only masters.
An explicit patch suffix is required for node version.

Latest "...specifies the highest supported Kubernetes version currently
available on GKE in the cluster's zone or region".