Ensure node-pool example node count < 6
When GKE stands up a cluster with more than 6 nodes in the node pool, it
will automatically resize the master node. This causes the cluster to
enter reconciling, which fouls up our CI configuration.

This commit sidesteps this problem by converting the node-pool example
to a zonal pool, removing the default node pool and shrinking the max
size of the added node pools. We can add graceful handling of cluster
reconciliation in a followup PR.