Add a `check_generate_docs` target
Currently we have no automatic way of determining if a given commit or
pull request has stale documentation; this leads to buildups in changes
that tend to tag along with other fixes.

This commit resolves the issue by adding a check that verifies that a
fresh copy of the repository does not have a diff present after docs
generation.