Add suffix to cluster service account
The service account ID truncation may drop appended to the cluster name;
this means that multiple clusters with a common prefix but different
name may try to create service accounts with identical IDs and then
fail.

This commit resolves the issue by adding a suffix to the service account
name to ensure uniqueness.