Creation of new variable configure_ip_masq controlling ip masq install
Aliased IP addresses do not require ip masquerading anymore.  There are  few use cases where we would
need ip masq, but usually it is not recommended to install ip masquerading. This variable allows
for fine gain control on the installation of ip masq as it was always installed via the
network_policy variable previously.

configure_ip_masq defaults to false.

Fixes: https://github.com/terraform-google-modules/terraform-google-kubernetes-engine/issues/172