Adding various beta features and code hygiene
This PR completes some basic code hygiene. We moved the database
encyption block in the same if statement as the other beta components.

With this commit we include various beta features that are supported at this time:

Cluster Configurations

- authenticator_groups_config
- vertical_pod_autoscaling
- workload_identity_config
- enable_intranode_visibility
- default_max_pods_per_node

We have TPU, some nodepool options, and other beta features that I have not put in this PR.
Will revist in later PRs.

This has been tested againt the 2.9.1 google provider.  Upgrading the examples to that
provider will occur in a different PR.