Add a parameter 'registry_project_id'
The PR allows configuring the project holding the GCR registry when
used in connection with 'create_service_account'=true and
grant_registry_access=true.

Holding the GCR is a project with other resources increases the risk of
exposing sensitive data to the service account running the nodes, as the
required permissions of role roles/storage.objectViewer provide access
to all storage objects in the project.