fix: incorrect `node_pools` variable type
`node_pool` variables are a mix of string, bool, number
therefore, it is better to use `map(any)` to indicate its type

moreover, the use of `list(map(string))` is breaking
the use of this module in `cdktf`.

in the generated code it will produce
`[]map[string]string` for `node_pools` option
but we want `[]map[string]interface{}` in order to pass correct value,
e.g. `initial_node_count` is a `number` not `string`