feat(cluster.tf): add support for setting initial version when using release channel
This change adds support for specifying an initial version to use when a release channel is being used.
The new logic sets the `min_master_version` variable to the value of the
locally defined `master_version` when `release_channel` is `null` or `"UNSPECIFIED"`.
When a release channel is specified, `min_master_version` is set to the value of `initial_version`
if it is defined, and falls back to the value of `master_version` otherwise.