Make `enable_binary_authorization` optional
By default `enable_binary_authorization` is set to `true` which means
that the Binary Authorization feature is enabled. This is the current
value of the variable in the `safer-cluster` module.

That feature makes sense only when it is used. Until it is not used,
it only takes money for nothing.

Fixes: https://github.com/terraform-google-modules/terraform-google-kubernetes-engine/issues/1632