fix: add `insecureKubeletReadonlyPortEnabled` to node_config
Add `insecureKubeletReadonlyPortEnabled` to `node_config.kubelet_config`
for the default node-pool and for additional pools. It may also be
necessary to define the top level `node_config` more broadly for the
case where `remove_default_node_pool` is set to false, which should
probably be handled separately.

Also, the upstream provider (intentionally) uses an enum of `"TRUE"` /
`"FALSE"` vs. a boolean. Update the code to follow this, and add a test
case that covers the cluster level setting vs node pool one.

Fixes #2013

Co-authored-by: Andrew Peabody <andrewpeabody@google.com>